[hotfix][tests] Move containsCause to FlinkMatchers
