+++ amazon-kinesis-connector-flink/src/test/java/software/amazon/kinesis/connectors/flink/table/KinesisDynamicTableFactoryTest.java
@@ -55,7 +55,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
-import static org.apache.flink.util.CoreMatchers.containsCause;
+import static org.apache.flink.core.testutils.FlinkMatchers.containsCause;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
